% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{run_all}
\alias{run_all}
\title{run_all}
\usage{
run_all(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
(invisible) list of results
}
\description{
Print all four results (A variance function in C++, B mean function in C++, mean in R, variance in R)
}
\examples{
run_all(c(1,2,3,4))
}
